feat(components/molecule/autosuggest): remove withOpenToogle HoC
