+++ components/molecule/autosuggest/demo/index.scss
@@ -3,6 +3,8 @@
 @import '~@s-ui/react-molecule-select/lib/index';
 @import '~@s-ui/react-atom-button/lib/index';
 
+@import '../src/index';
+
 .DemoMoleculeAutosuggest {
   margin-bottom: 500px;
   padding: 20px;
